+++ packages/ui/core-components/src/lib/atoms/inputs/dropdown/stories/DependentDropdowns.story.svelte
@@ -3,7 +3,6 @@
 	import { query } from '@evidence-dev/universal-sql/client-duckdb';
 	import Dropdown from '../Dropdown.svelte';
 	import DropdownOption from '../helpers/DropdownOption.svelte';
-	import QueryLoad from '../../../query-load/QueryLoad.svelte';
 	import { getContext } from 'svelte';
 	import { INPUTS_CONTEXT_KEY } from '@evidence-dev/component-utilities/globalContexts';
 	import { browser } from '$app/environment';
@@ -41,9 +40,13 @@
 	});
 </script>
 
-<Dropdown defaultValue={0} name="hashtag" data={baseQuery} value="value" label="label" />
-<QueryLoad let:loaded data={depQuery}>
-	<Dropdown name="user" data={loaded} value="value" label="label" multiple selectAllByDefault>
-		<DropdownOption value="All" />
-	</Dropdown>
-</QueryLoad>
+<div class="grid grid-cols-2 gap-8">
+	<div>
+		<Dropdown defaultValue={0} name="hashtag" data={baseQuery} value="value" label="label" />
+	</div>
+	<div>
+		<Dropdown name="user" data={depQuery} value="value" label="label" multiple selectAllByDefault>
+			<DropdownOption value="All" />
+		</Dropdown>
+	</div>
+</div>